Support for changing the startup type on the ArcGIS Server
This is a minimal implementation for allowing the configuration of the ArcGIS Server service startup type. It does not allow the alteration of geoevent components nor the notebook server.

A more global implementation may consider replacing/altering the `SetStartupToAutomatic` parameter and affecting any of the DSC-controlled components.

This is intended as a use-at-your-own-risk /  YMMV / quality of life change for deployments facing CPU resource constraints.

See: <https://github.com/Esri/arcgis-powershell-dsc/issues/609>